Origin: Accuracy adjustments, powered staffs non-attackable in pvp, dinhs bulwark can no longer attack players in defensive state

+++ src/main/java/com/aelous/model/entity/combat/formula/accuracy/RangeAccuracy.java
@@ -53,7 +53,7 @@ public class RangeAccuracy {
             successfulRoll = attackBonus / (2F * (defenceBonus + 1F));
         }
 
-        double selectedChance = random.nextDouble();
+        double selectedChance = random.nextFloat();
 
         System.out.println("PlayerStats - Attack=" + attackBonus + " Def=" + defenceBonus + " chanceOfSucess=" + new DecimalFormat("0.000").format(successfulRoll) + " rolledChance=" + new DecimalFormat("0.000").format(selectedChance) + " successful=" + (successfulRoll > selectedChance ? "YES" : "NO"));
